+++ src/components/GameCard/GameCard.module.css
@@ -3,21 +3,23 @@
   justify-content: space-between;
   align-items: center;
   padding: 16px;
-  border: 1px solid #444;
+  border: 1px solid #3a3a5e; 
   border-radius: 8px;
-  background-color: #2f2f2f;
+  background-color: #2f2f4f; 
   transition: transform 0.2s ease-in-out, box-shadow 0.2s ease-in-out;
 }
 
 .card:hover {
   transform: translateY(-4px);
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.4);
+  border-color: #646cff;
 }
 
 .gameName {
   font-size: 1.2rem;
   font-weight: bold;
   margin: 0;
+  color: #f0f0f0;
 }
 
 .details {
@@ -38,4 +40,22 @@
 
 .joinButton:hover {
   background-color: #535bf2;
+}
+
+.cardAnimated {
+  animation: fadeInUp 0.5s ease-out forwards;
+  animation-fill-mode: backwards;
+}
+
+
+
+@keyframes fadeInUp {
+  from {
+    opacity: 0;
+    transform: translateY(20px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
 }
