+++ src/containers/App/App.css
@@ -1,6 +1,6 @@
-.App-container {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  text-align: center;
+body {
+  margin: 0;
+  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  background-color: #1a1a2e; 
+  color: #e0e0e0; 
 }
