+++ src/containers/App/App.jsx
@@ -2,7 +2,8 @@
 
 import React from 'react';
 import './App.css';
-import HomePage from '@/pages/HomePage.jsx';
+import HomePage from '@/pages/HomePage/HomePage.jsx';
+
 
 function App() {
   return (
