+++ src/pages/HomePage/HomePage.module.css
@@ -1,0 +1,53 @@
+.pageContainer {
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  align-items: center;   
+  padding: 2rem;
+  box-sizing: border-box;
+
+
+  background-image: url('@/pages/HomePage/Gemini_Generated_Image_5h35hi5h35hi5h35.png'); 
+  background-position: center;  
+  background-repeat: no-repeat; 
+  background-attachment: fixed; 
+
+}
+
+.contentWrapper {
+  width: 100%;
+  max-width: 800px; 
+  display: flex;
+  flex-direction: column;
+}
+
+
+
+.header {
+  display: flex;
+  justify-content: space-between; 
+  align-items: center;
+  width: 100%;
+  margin-bottom: 2rem;
+}
+
+.title {
+  margin: 0;
+  color: #c7c7ff;
+}
+
+.createButton {
+  background-color: #3fa844;
+  color: white;
+  border: none;
+  padding: 10px 20px;
+  border-radius: 5px;
+  cursor: pointer;
+  font-size: 1rem;
+  font-weight: bold;
+  transition: background-color 0.2s;
+}
+
+.createButton:hover {
+  background-color: #358d3a;
+}
